+++ src/index.js
@@ -1,2 +1,5 @@
 // @flow
-export { default, DropzoneWithoutContext } from './Dropzone';
+export { default } from './Dropzone';
+
+export { default as Target } from './Target';
+export { default as DropTarget } from './DropTarget';
